Remove legacy apps, environments and schedules
* Moved files from client.api.teams.File to client.api.file.File
* File documentation removed from api.__init__ docstring pending further work
