--- projects/python-client/src/datapane/client/apps/__init__.py
@@ -1,8 +1,0 @@
-from .builder import build_bundle
-from .config import DATAPANE_YAML, DatapaneCfg
-
-__all__ = [
-    "build_bundle",
-    "DATAPANE_YAML",
-    "DatapaneCfg",
-]
