+++ projects/python-client/src/datapane/client/utils.py
@@ -1,14 +1,11 @@
-import argparse
 import os
 import string
 import sys
 import typing as t
-from distutils.util import strtobool
-from typing import Iterator, Tuple
 
 import click
 
-from datapane.common import DPError, JDict
+from datapane.common import DPError
 
 
 ################################################################################
@@ -113,60 +110,3 @@ def display_msg(text: str, **params: str):
         display(Markdown(msg))
     else:
         print(msg)
-
-
-################################################################################
-# Misc
-def process_cmd_param_vals(params: Tuple[str, ...]) -> JDict:
-    """Convert a list of k=v to a typed JSON dict"""
-
-    def convert_param_val(x: str) -> t.Union[int, float, str, bool]:
-        # TODO - this can be optimised / cleaned-up
-        try:
-            return int(x)
-        except ValueError:
-            try:
-                return float(x)
-            except ValueError:
-                try:
-                    return bool(strtobool(x))
-                except ValueError:
-                    return x
-
-    def split_param(xs: Tuple[str, ...]) -> Iterator[t.Tuple[str, str]]:
-        err_msg = "'{}', should be name=value"
-        for x in xs:
-            try:
-                k, v = x.split("=", maxsplit=1)
-            except ValueError:
-                raise click.BadParameter(err_msg.format(x))
-            if not k or not v:
-                raise click.BadParameter(err_msg.format(x))
-            yield (k, v)
-
-    return {k: convert_param_val(v) for (k, v) in split_param(params)}
-
-
-def parse_command_line() -> t.Dict[str, t.Any]:
-    """Called in library mode to read any datapane CLI parameters"""
-    parser = argparse.ArgumentParser(
-        description="Datapane additional args",
-        conflict_handler="resolve",
-        add_help=False,
-    )
-    parser.add_argument(
-        "--parameter",
-        "-p",
-        action="append",
-        help="key=value parameters to pass into dp.Params",
-        default=[],
-    )
-
-    (dp_args, remaining_args) = parser.parse_known_args()
-    # reset sys.argv without the dp params for parsing by the user
-    exe_name = sys.argv.pop(0)
-    sys.argv.clear()
-    sys.argv.append(exe_name)
-    sys.argv.extend(remaining_args)
-
-    return process_cmd_param_vals(dp_args.parameter)
