--- projects/python-client/tests/client/apps/test_builder/README.md
@@ -1,9 +1,0 @@
-# test_proj1
-
-_TODO_ - describe app
-
-Add GIT support
-
-## Config
-
-This app can be configured using the following parameters...
