--- projects/python-client/tests/client/apps/test_builder/config.yaml
@@ -1,25 +1,0 @@
-# App name
-name: Test proj 2
-
-# App parameter definitions - see docs.data
-# see xyz for configuration docs
-parameters: [
-  {
-    "type": "string",
-    "name": "testStr1",
-  },
-  {
-    "type": "integer",
-    "name": "testInt",
-    "default": 4
-  }
-]
-
-# CURRENTLY UNUSED
-# additional_fields: True
-# language: python
-# docker_image: datapane-python-default
-# volatile: False
-# privileged: False
-# Persist the app between requests
-# persistent: True
