--- projects/python-client/tests/client/apps/test_config/dp_app.ipynb.yaml
@@ -1,14 +1,0 @@
-name: dp_test_app
-
-parameters: [
-  {"name": "param1", "type": "string", "default": "hello"},
-  {"name": "param2", "type": "integer", "required": true}
-]
-
-source_url: https://github.com/datapane/test
-
-include:
-  - sample.ipynb
-
-requirements:
-  - requests
