--- projects/python-client/tests/client/apps/test_config/dp_app.py.yaml
@@ -1,12 +1,0 @@
-name: dp_test_app
-
-parameters: [
-  {"name": "param1", "type": "string", "default": "hello"},
-  {"name": "param2", "type": "integer", "required": true}
-]
-
-include:
-  - sample.py
-
-requirements:
-  - requests
