+++ projects/python-client/tests/client/e2e/test_cli.py
@@ -1,12 +1,9 @@
-from pathlib import Path
-
 import pytest
 from click.testing import CliRunner
 
 from datapane.client.commands import cli
 
 from ..local.test_cli import handle_res
-from .common import gen_name
 from .conftest import TEST_SERVER, TEST_TOKEN
 
 
@@ -30,23 +27,3 @@ def test_auth(runner: CliRunner):
     result = runner.invoke(cli, ["-vv", "logout"])
     handle_res(result)
     assert "Logged out" in result.output
-
-
-@pytest.mark.skip("LeagacyApp tests disabled")
-@pytest.mark.org
-@pytest.mark.timeout(5 * 60)  # allow 5m
-def test_cli_app(runner: CliRunner, tmp_path: Path, monkeypatch):
-    monkeypatch.chdir(tmp_path)
-
-    r = runner.invoke(cli, ["-vv", "app", "init"])
-    handle_res(r)
-    app_name = gen_name()
-    r = runner.invoke(cli, ["-vv", "app", "deploy", "--name", app_name])
-    handle_res(r)
-
-    if "Uploaded" in r.output:
-        r = runner.invoke(cli, ["-vv", "app", "run", app_name, "--wait"])
-        handle_res(r)
-
-        r = runner.invoke(cli, ["-vv", "app", "delete", app_name])
-        handle_res(r)
