+++ projects/python-client/tests/client/local/test_cli.py
@@ -3,7 +3,6 @@ from pathlib import Path
 from click.testing import CliRunner, Result
 
 from datapane.client.commands import cli
-from datapane.client.utils import process_cmd_param_vals
 
 # TODO
 #  - add tests for other commands
@@ -24,17 +23,3 @@ def test_report_init(tmp_path: Path, monkeypatch):
     runner = CliRunner()
     result = runner.invoke(cli, ["-vv", "report", "init"])
     handle_res(result)
-
-
-def test_app_init(tmp_path: Path, monkeypatch):
-    monkeypatch.chdir(tmp_path)
-
-    runner = CliRunner()
-    result = runner.invoke(cli, ["-vv", "app", "init"])
-    handle_res(result)
-
-
-def test_cli_params():
-    p1 = ("w=hello", "x=1", "y=3.2", "z=true")
-    expected1 = {"w": "hello", "x": 1, "y": 3.2, "z": True}
-    assert process_cmd_param_vals(p1) == expected1
