--- projects/python-client/tests/runner/test_runner
@@ -1,1 +1,0 @@
-../testfiles/runscripts
