--- projects/python-client/tests/testfiles/runscripts/datapane.yaml
@@ -1,4 +1,0 @@
-name: dp Test app
-script: dp_app.py
-parameters:
-  - {"name":"p1", "type": "string"}
